+++ idm_heatpump/__init__.py
@@ -8,6 +8,7 @@ from .client import (
     RegisterDef,
     RegisterType,
     WriteClass,
+    quiet_pymodbus_logging,
 )
 from .const import (
     ACTIVE_HC_MODE_OPTIONS,
@@ -139,5 +140,6 @@ __all__ = [
     "get_heating_circuit_registers",
     "get_register",
     "get_zone_module_registers",
+    "quiet_pymodbus_logging",
     "web_pin_configured",
 ]